split body into req and res
