+++ elastic/examples/async_sample/src/body/request.rs
@@ -1,14 +1,13 @@
-use std::cmp;
 use std::collections::VecDeque;
 use std::path::Path;
-use std::io::{Read, Error as IoError};
+use std::io::Error as IoError;
 
 use tokio_proto::streaming::Body as BodyStream;
 use futures::{IntoFuture, Future, Stream, Sink, Poll, Async};
 use futures::future::lazy;
 use futures::sync::mpsc::SendError;
 use memmap::{Mmap, MmapViewSync, Protection};
-use hyper::{Chunk as HyperChunk, Error as HyperError};
+use hyper::Error as HyperError;
 
 use error::*;
 
@@ -75,12 +74,7 @@ pub fn mapped_file<P>
 fn map_file_to_chunks<P>(path: P) -> Result<VecDeque<FileChunkResult>, Error>
     where P: AsRef<Path>
 {
-    let file = match Mmap::open_path(path, Protection::Read) {
-        Ok(file) => file,
-        Err(e) => return Err(e.into()),
-    };
-
-    let file = file.into_view_sync();
+    let file = Mmap::open_path(path, Protection::Read)?.into_view_sync();
 
     let total_len = file.len();
     if total_len == 0 {
@@ -95,10 +89,7 @@ fn map_file_to_chunks<P>(path: P) -> Result<VecDeque<FileChunkResult>, Error>
         next = match rest.len() {
             // >1 chunk size left
             len if len > SLICE_SIZE => {
-                let (slice, rest) = match rest.split_at(SLICE_SIZE) {
-                    Ok(split) => split,
-                    Err(e) => return Err(e.into()),
-                };
+                let (slice, rest) = rest.split_at(SLICE_SIZE)?;
 
                 slices.push_back(Ok(FileChunk(slice)));
 
@@ -116,87 +107,4 @@ fn map_file_to_chunks<P>(path: P) -> Result<VecDeque<FileChunkResult>, Error>
     }
 
     Ok(slices)
-}
-
-/// A readable wrapper around a `Chunk`.
-///
-/// This type is basically the same as `std::io::Cursor`.
-pub struct Chunk {
-    buf: HyperChunk,
-    len: usize,
-    pos: usize,
-}
-
-impl From<HyperChunk> for Chunk {
-    fn from(chunk: HyperChunk) -> Self {
-        Chunk {
-            len: chunk.len(),
-            pos: 0,
-            buf: chunk,
-        }
-    }
-}
-
-impl Read for Chunk {
-    #[inline]
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, IoError> {
-        let amt = cmp::min(self.pos, self.len);
-        let mut chunk = &self.buf[amt..];
-
-        let read = chunk.read(buf)?;
-        self.pos += read;
-
-        Ok(read)
-    }
-}
-
-type ChunkSequence = VecDeque<Chunk>;
-
-/// A builder for a sequence of chunks.
-pub struct ChunkBodyBuilder(ChunkSequence);
-
-/// A readable response body where bytes are read from all chunks without exposing them directly.
-pub struct ChunkBody(ChunkSequence);
-
-impl ChunkBodyBuilder {
-    pub fn new() -> Self {
-        ChunkBodyBuilder(ChunkSequence::new())
-    }
-
-    pub fn append<I>(&mut self, chunk: I)
-        where I: Into<Chunk>
-    {
-        self.0.push_back(chunk.into());
-    }
-
-    pub fn build(self) -> ChunkBody {
-        let mut chunks = self.0;
-        chunks.shrink_to_fit();
-
-        ChunkBody(chunks)
-    }
-}
-
-impl Read for ChunkBody {
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize, IoError> {
-        let mut pop = false;
-
-        let read = if let Some(mut chunk) = self.0.front_mut() {
-            let read = chunk.read(buf)?;
-
-            if chunk.pos >= chunk.len {
-                pop = true;
-            }
-
-            read
-        } else {
-            0
-        };
-
-        if pop {
-            self.0.pop_front();
-        }
-
-        Ok(read)
-    }
 }
