+++ elastic/examples/async_sample/src/main.rs
@@ -33,7 +33,7 @@ use hyper::client::HttpConnector;
 
 use error::Error;
 
-type RequestBody = body::FileBody;
+type RequestBody = body::request::FileBody;
 
 type AllocatedField = string_cache::DefaultAtom;
 type BulkResponse = elastic_responses::BulkErrorsResponse<AllocatedField, AllocatedField, String>;
@@ -64,21 +64,21 @@ fn send_request(url: &'static str,
                 pool: CpuPool)
                 -> impl Future<Item = BulkResponse, Error = Error> {
     // Get a future to buffer a bulk file
-    let (buffer_request_body, request_body) = body::mapped_file("./data/accounts.json").unwrap();
+    let (buffer_request_body, request_body) = body::request::mapped_file("./data/accounts.json").unwrap();
     let buffer_request_body = pool.spawn(buffer_request_body);
 
     // Build a Bulk request
-    let req = BulkRequest::for_index_ty("bulk-async", "bulk-ty", request_body);
+    let request = BulkRequest::for_index_ty("bulk-async", "bulk-ty", request_body);
 
     // Send the request
     let send_request = client
-        .request(hyper_req::build(&url, req))
+        .request(hyper_req::build(&url, request))
         .map_err(Into::into);
 
     // Buffer the response chunks into a synchronously readable sequence
     let read_response = send_request.and_then(move |response| {
         let status: u16 = response.status().into();
-        let chunks = body::ChunkBodyBuilder::new();
+        let chunks = body::response::ChunkBodyBuilder::new();
 
         response.body()
             .fold(chunks, concat_chunks)
@@ -101,9 +101,9 @@ fn send_request(url: &'static str,
         .map(move |(_, response)| response)
 }
 
-fn concat_chunks(mut chunks: body::ChunkBodyBuilder,
+fn concat_chunks(mut chunks: body::response::ChunkBodyBuilder,
                  chunk: hyper::Chunk)
-                 -> Result<body::ChunkBodyBuilder, hyper::Error> {
+                 -> Result<body::response::ChunkBodyBuilder, hyper::Error> {
     chunks.append(chunk);
     Ok(chunks)
 }
